Moved SharedObjectSequence to the deprecated package
