+++ api-report/deprecated.api.md
@@ -10,12 +10,17 @@ import { IChannelServices } from '@fluidframework/datastore-definitions';
 import { IChannelStorageService } from '@fluidframework/datastore-definitions';
 import { IFluidDataStoreRuntime } from '@fluidframework/datastore-definitions';
 import { IFluidSerializer } from '@fluidframework/shared-object-base';
+import { IJSONSegment } from '@fluidframework/merge-tree';
 import { Interval } from '@fluidframework/sequence';
 import { IntervalCollection } from '@fluidframework/sequence';
 import { ISequencedDocumentMessage } from '@fluidframework/protocol-definitions';
 import { ISharedIntervalCollection } from '@fluidframework/sequence';
+import { ISharedObject } from '@fluidframework/shared-object-base';
 import { ISummaryTreeWithStats } from '@fluidframework/runtime-definitions';
+import { Serializable } from '@fluidframework/datastore-definitions';
 import { SharedObject } from '@fluidframework/shared-object-base';
+import { SharedSequence } from '@fluidframework/sequence';
+import { SubSequence } from '@fluidframework/sequence';
 
 // @public @deprecated (undocumented)
 export class SharedIntervalCollection extends SharedObject implements ISharedIntervalCollection<Interval> {
@@ -56,6 +61,38 @@ export class SharedIntervalCollectionFactory implements IChannelFactory {
     // (undocumented)
     static readonly Type = "https://graph.microsoft.com/types/sharedIntervalCollection";
     // (undocumented)
+    get type(): string;
+}
+
+// @public @deprecated
+export class SharedObjectSequence<T> extends SharedSequence<T> {
+    // @deprecated
+    constructor(document: IFluidDataStoreRuntime, id: string, attributes: IChannelAttributes);
+    // @deprecated
+    static create<T>(runtime: IFluidDataStoreRuntime, id?: string): SharedObjectSequence<T>;
+    // @deprecated
+    static getFactory(): SharedObjectSequenceFactory;
+    // @deprecated (undocumented)
+    getRange(start: number, end?: number): Serializable<T>[];
+    // (undocumented)
+    id: string;
+}
+
+// @public @deprecated (undocumented)
+export class SharedObjectSequenceFactory implements IChannelFactory {
+    // @deprecated (undocumented)
+    static readonly Attributes: IChannelAttributes;
+    // @deprecated (undocumented)
+    get attributes(): IChannelAttributes;
+    // @deprecated (undocumented)
+    create(document: IFluidDataStoreRuntime, id: string): ISharedObject;
+    // @deprecated (undocumented)
+    load(runtime: IFluidDataStoreRuntime, id: string, services: IChannelServices, attributes: IChannelAttributes): Promise<ISharedObject>;
+    // @deprecated (undocumented)
+    static segmentFromSpec(segSpec: IJSONSegment): SubSequence<object>;
+    // @deprecated (undocumented)
+    static Type: string;
+    // @deprecated (undocumented)
     get type(): string;
 }
