+++ api-report/sequence.api.md
@@ -471,38 +471,6 @@ export class SharedNumberSequenceFactory implements IChannelFactory {
     get type(): string;
 }
 
-// @public @deprecated
-export class SharedObjectSequence<T> extends SharedSequence<T> {
-    // @deprecated
-    constructor(document: IFluidDataStoreRuntime, id: string, attributes: IChannelAttributes);
-    // @deprecated
-    static create<T>(runtime: IFluidDataStoreRuntime, id?: string): SharedObjectSequence<T>;
-    // @deprecated
-    static getFactory(): SharedObjectSequenceFactory;
-    // @deprecated (undocumented)
-    getRange(start: number, end?: number): Serializable<T>[];
-    // (undocumented)
-    id: string;
-}
-
-// @public @deprecated (undocumented)
-export class SharedObjectSequenceFactory implements IChannelFactory {
-    // @deprecated (undocumented)
-    static readonly Attributes: IChannelAttributes;
-    // @deprecated (undocumented)
-    get attributes(): IChannelAttributes;
-    // @deprecated (undocumented)
-    create(document: IFluidDataStoreRuntime, id: string): ISharedObject;
-    // @deprecated (undocumented)
-    load(runtime: IFluidDataStoreRuntime, id: string, services: IChannelServices, attributes: IChannelAttributes): Promise<ISharedObject>;
-    // @deprecated (undocumented)
-    static segmentFromSpec(segSpec: IJSONSegment): SubSequence<object>;
-    // @deprecated (undocumented)
-    static Type: string;
-    // @deprecated (undocumented)
-    get type(): string;
-}
-
 // @public (undocumented)
 export abstract class SharedSegmentSequence<T extends ISegment> extends SharedObject<ISharedSegmentSequenceEvents> implements ISharedIntervalCollection<SequenceInterval> {
     constructor(dataStoreRuntime: IFluidDataStoreRuntime, id: string, attributes: IChannelAttributes, segmentFromSpec: (spec: IJSONSegment) => ISegment);
