+++ experimental/dds/deprecated/src/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from "./sharedIntervalCollection";
+export * from "./sharedObjectSequence";
