+++ packages/dds/sequence/package.json
@@ -121,6 +121,14 @@
       "RemovedClassDeclaration_SharedIntervalCollectionFactory": {
         "backCompat": false,
         "forwardCompat": false
+      },
+      "RemovedClassDeclaration_SharedObjectSequence": {
+        "backCompat": false,
+        "forwardCompat": false
+      },
+      "RemovedClassDeclaration_SharedObjectSequenceFactory": {
+        "backCompat": false,
+        "forwardCompat": false
       }
     }
   }
