+++ packages/tools/replay-tool/src/replayFluidFactories.ts
@@ -29,11 +29,13 @@ import {
 import { create404Response, RuntimeFactoryHelper } from "@fluidframework/runtime-utils";
 import {
     SharedNumberSequence,
-    SharedObjectSequence,
     SharedString,
     SparseMatrix,
 } from "@fluidframework/sequence";
-import { SharedIntervalCollection } from "@fluid-experimental/deprecated";
+import {
+    SharedIntervalCollection,
+    SharedObjectSequence,
+} from "@fluid-experimental/deprecated";
 import { SharedSummaryBlock } from "@fluidframework/shared-summary-block";
 import { UnknownChannelFactory } from "./unknownChannel";
